Move ops for nodeDNS to nodedns package
